pyproject.toml
[tool.poetry]
name = "blackhat-python"
version = "0.1.0"
description = "Tools from Black Hat Python"
authors = ["Lucas Melin <lucas.melin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10"
scapy = "^2.4.5"
paramiko = "^2.8.1"
opencv-python = "^4.5.4"
requests = "^2.26.0"
lxml = "^4.7.1"
"github3.py" = "^3.0.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
